// src/app/modules/reports/report.interface.ts
export type TReport = {
  img_tagline_bangla: string;
  img_tagline_english: string;
  admin_name: string;
  date: string;
  bangla_title: string;
  english_title: string;
  category: string;
  bangla_short_description: string;
  english_short_description: string;
  bangla_description: string;
  english_description: string;
  name_published_newspaper: string;
  news_release_date: string;
  Link_published_newspaper: string;
  published_newspapers: {
    name: string;
    release_date: string;
    link: string;
    screenshot: string[];
  }[];
  meta_title: string;
  meta_keywords: string[];
  meta_description: string;
  thumnail_img: string;
  img_english: string;
  img_bangla: string;
  slug: string;
  eng_iamges: string[];
  bng_Images: string[];
  newsCategory: string;
};